refactor: move temporal-detection types/predicates to @fiftyone/utilities
