+++ app/packages/utilities/src/index.ts
@@ -20,6 +20,7 @@ export * from "./Resource";
 export * from "./schema";
 export { default as sizeBytesEstimate } from "./size-bytes-estimate";
 export * as styles from "./styles";
+export * from "./temporal";
 export * from "./type-check";
 export * from "./types";
 export * from "./validation";
